requirements.txt
numpy
scipy
soundfile
# Video pipeline (place.py + render.py)
pillow
imageio-ffmpeg
